refactor: :recycle: update delete database method

+++ api/database/delete.go
@@ -7,10 +7,10 @@ import (
 	"go.mongodb.org/mongo-driver/bson"
 )
 
-func (mdb *MongoDatabase) Delete(taskId uuid.UUID) error {
+func (mdb *MongoDatabase) Delete(userId, taskId uuid.UUID) error {
 	coll := mdb.client.Database("pair-challenge").Collection("tasks")
 
-	res, err := coll.DeleteOne(context.Background(), bson.M{"_id": taskId})
+	res, err := coll.DeleteOne(context.Background(), bson.M{"_id": taskId, "userId": userId})
 	if err != nil {
 		mdb.logger.Errorf("database.Delete: mongo.DeleteOne >> %v", err)
 		return err
